moved training code outside of main function
