+++ train_config.json
@@ -5,7 +5,7 @@
     "batch_size": 1024,
     "val_batch_size": 4096,
     "lr": 1e-3,
-    "gpu": [4],
+    "gpu": [7],
     "epochs": 1,
     "save_model": "22052025_Focal_focalalpha0_8_focalgamma_1_alpha5000_beta0_gamma_0_epoch300_eta10_hubermargin10_5channel_400bps_model.pth",
     "log_interval": 10,
